#8 Update
Still having the issue where the dms wont work but iv found the cause. The bot cant send a DM to the user if the bot was turned on while the user was still in the server then kicked, however if the user isnt in the server when the bot tries to DM them it works fine.